src/main/java/org/learning/hackerrank/MatchingStrings.java
package org.learning.hackerrank;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class MatchingStrings {

    public static List<Integer> matchingStrings(List<String> strings, List<String> queries) {
        List<Integer> result = new ArrayList<>();
        Map<String, Integer> map = new HashMap<>();
        for (String s : strings) {
            if (map.containsKey(s)) {
                map.put(s, map.get(s) + 1);
            } else {
                map.put(s, 1);
            }
        }
        for (String s : queries) {
            if(!map.containsKey(s)){
                result.add(0);
            }else{
                result.add(map.get(s));
            }
        }
        return result;
    }
}
